specific app stats and popular apps
changed up the popular apps tab a lot. now in the tab you can search for any app and it gives you suggestions what apps are in the database (probably buggy). By searching you can see specific stats of any app or you can press on a row card of one of the popular apps to see its stats.